style: move clusterinstall to util pkg
