+++ pkg/core/manager/cluster/manager.go
@@ -21,8 +21,8 @@ import (
 	"time"
 
 	clusterv1beta1 "github.com/KusionStack/karbour/pkg/apis/cluster/v1beta1"
-	"github.com/KusionStack/karbour/pkg/clusterinstall"
 	"github.com/KusionStack/karbour/pkg/multicluster"
+	"github.com/KusionStack/karbour/pkg/util/clusterinstall"
 	"github.com/KusionStack/karbour/pkg/util/ctxutil"
 	yaml "gopkg.in/yaml.v3"
 	v1 "k8s.io/api/core/v1"
